Move tag, comment into post domain
