+++ src/main/java/com/blind/company/domain/CompanyCategory.java
@@ -1,21 +1,19 @@
-package com.blind.api.tag.domain;
+package com.blind.company.domain;
 
-import com.blind.common.domain.BaseEntity;
-import lombok.AllArgsConstructor;
+import com.blind.shared.domain.BaseEntity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EntityListeners;
 
 @Entity
 @Getter
-@Setter
 @NoArgsConstructor
-@AllArgsConstructor
 @EntityListeners(AuditingEntityListener.class)
-public class Tag extends BaseEntity {
-    private String tagName;
+public class CompanyCategory extends BaseEntity {
+    @Column(nullable = false)
+    private String name;
 }
