+++ src/main/java/com/blind/company/persistence/respository/CompanyCategoryRepository.java
@@ -1,6 +1,6 @@
 package com.blind.company.persistence.respository;
 
-import com.blind.company.domain.Entity.CompanyCategory;
+import com.blind.company.domain.CompanyCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
