+++ src/main/java/com/blind/company/persistence/respository/JobCategoryRepository.java
@@ -1,6 +1,6 @@
 package com.blind.company.persistence.respository;
 
-import com.blind.company.domain.Entity.JobCategory;
+import com.blind.company.domain.JobCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
