+++ src/main/java/com/blind/company/persistence/respository/JobPositionRepository.java
@@ -1,6 +1,6 @@
 package com.blind.company.persistence.respository;
 
-import com.blind.company.domain.Entity.JobPosition;
+import com.blind.company.domain.JobPosition;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
