+++ src/main/java/com/blind/post/domain/Tag.java
@@ -1,19 +1,21 @@
-package com.blind.company.domain.Entity;
+package com.blind.post.domain;
 
-import com.blind.common.domain.BaseEntity;
+import com.blind.shared.domain.BaseEntity;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EntityListeners;
 
 @Entity
 @Getter
+@Setter
 @NoArgsConstructor
+@AllArgsConstructor
 @EntityListeners(AuditingEntityListener.class)
-public class CompanyCategory extends BaseEntity {
-    @Column(nullable = false)
-    private String name;
+public class Tag extends BaseEntity {
+    private String tagName;
 }
