+++ src/main/java/com/blind/post/persistence/repository/BookmarkRepository.java
@@ -1,6 +1,6 @@
 package com.blind.post.persistence.repository;
 
-import com.blind.post.domain.Entity.Bookmark;
+import com.blind.post.domain.Bookmark;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
