+++ src/main/java/com/blind/post/persistence/repository/TagRepository.java
@@ -1,6 +1,6 @@
-package com.blind.api.tag.persistence;
+package com.blind.post.persistence.repository;
 
-import com.blind.api.tag.domain.Tag;
+import com.blind.post.domain.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagRepository extends JpaRepository<Tag,Integer> {
